refactor lambda starting code
