+++ cmd/lambda/notifier/main.go
@@ -5,42 +5,19 @@ import (
 	"time"
 
 	"github.com/aws/aws-lambda-go/events"
-	"github.com/aws/aws-lambda-go/lambda"
 	logging "github.com/ipfs/go-log/v2"
+	"github.com/storacha/indexing-service/cmd/lambda"
 	"github.com/storacha/indexing-service/pkg/aws"
-	"github.com/storacha/indexing-service/pkg/telemetry"
 	"github.com/storacha/ipni-publisher/pkg/notifier"
-	"go.opentelemetry.io/contrib/instrumentation/github.com/aws/aws-lambda-go/otellambda"
 )
 
 var log = logging.Logger("lambda/notifier")
 
 func main() {
-	cfg := aws.FromEnv(context.Background())
-
-	// an empty API key disables instrumentation
-	if cfg.HoneycombAPIKey != "" {
-		ctx := context.Background()
-		tp, telemetryShutdown, err := telemetry.SetupTelemetry(ctx, &cfg.Config)
-		if err != nil {
-			panic(err)
-		}
-		defer telemetryShutdown(ctx)
-
-		handler := makeHandler(cfg)
-
-		instrumentedHandler := otellambda.InstrumentHandler(
-			handler,
-			otellambda.WithTracerProvider(tp),
-			otellambda.WithFlusher(tp),
-		)
-		lambda.StartWithOptions(instrumentedHandler, lambda.WithContext(ctx))
-	} else {
-		lambda.Start(makeHandler(cfg))
-	}
+	lambda.Start(makeHandler)
 }
 
-func makeHandler(cfg aws.Config) func(ctx context.Context, event events.EventBridgeEvent) {
+func makeHandler(cfg aws.Config) any {
 	// setup IPNI
 	// TODO: switch to double hashed client for reader privacy?
 	headStore := aws.NewS3Store(cfg.Config, cfg.NotifierHeadBucket, "")
