+++ cmd/lambda/remotesync/main.go
@@ -5,45 +5,23 @@ import (
 	"encoding/json"
 
 	"github.com/aws/aws-lambda-go/events"
-	"github.com/aws/aws-lambda-go/lambda"
 	"github.com/ipfs/go-cid"
 	cidlink "github.com/ipld/go-ipld-prime/linking/cid"
 	goredis "github.com/redis/go-redis/v9"
 	"github.com/storacha/go-metadata"
+	"github.com/storacha/indexing-service/cmd/lambda"
 	"github.com/storacha/indexing-service/pkg/aws"
 	"github.com/storacha/indexing-service/pkg/redis"
 	"github.com/storacha/indexing-service/pkg/service/providerindex"
 	"github.com/storacha/indexing-service/pkg/telemetry"
 	"github.com/storacha/ipni-publisher/pkg/store"
-	"go.opentelemetry.io/contrib/instrumentation/github.com/aws/aws-lambda-go/otellambda"
 )
 
 func main() {
-	cfg := aws.FromEnv(context.Background())
-
-	// an empty API key disables instrumentation
-	if cfg.HoneycombAPIKey != "" {
-		ctx := context.Background()
-		tp, telemetryShutdown, err := telemetry.SetupTelemetry(ctx, &cfg.Config)
-		if err != nil {
-			panic(err)
-		}
-		defer telemetryShutdown(ctx)
-
-		handler := makeHandler(cfg)
-
-		instrumentedHandler := otellambda.InstrumentHandler(
-			handler,
-			otellambda.WithTracerProvider(tp),
-			otellambda.WithFlusher(tp),
-		)
-		lambda.StartWithOptions(instrumentedHandler, lambda.WithContext(ctx))
-	} else {
-		lambda.Start(makeHandler(cfg))
-	}
+	lambda.Start(makeHandler)
 }
 
-func makeHandler(cfg aws.Config) func(ctx context.Context, snsEvent events.SNSEvent) error {
+func makeHandler(cfg aws.Config) any {
 	var providersRedis *goredis.Client
 	if cfg.HoneycombAPIKey != "" {
 		providersRedis = telemetry.GetInstrumentedRedisClient(&cfg.ProvidersRedis)
